shortened statement that determined the value of cutout passed to get_uri.
The 'locations' array holds a list of places that the get_image method will iterate through trying to find our image for us.  The list is a set of tuples.   (uri, cutout) where cutout (if not None) is passed as an argument to the data service.

For reasons that are lost on me the value set in the tuple was:

cutout is not None and cutout or None

So.. if the variable cutout is not None then the value in the variable is sent, otherwise the value None is sent.  But the value in cutout is None so basically we always send the value of the variable and the logic is just extra compute steps?

Why was this written this way before?