HACK to set the value of 'is_inverted'.
The flag 'is_inverted' is set to TRUE when we notice the the x/y coordinates of storage locations is flip/flopped compared to the orignal image.  This flip/flopping is done when removing a single extension from an MEF to ensure that North is up and East is left when passing that to OSSOS detection pipeline since the pipeline works in X/Y space not RA/DEC space.

 Why is this in astrom you ask.   The parse reads the measurements from a pipeline output and needs to know if it should try and 'flip/flop' the images before matching the recorded location to the image location.  The original MEF files, which are split for running the pipeline, are always stored in the MEF and thus must be flip/flopped but the 'scrambled' and 'fk' images are NOT stored in teh MEF but rather was single images... so they must not be flip/flopped.

  is_inverted === is the coordinate I just recorded flip/flop compared to how the image is stored on disk?

  One day we'll work in RA/DEC and this problem will go away... that day will never cover I suspect.

  @mtb55: This is to handle an original pipeline feature.