Moved default parameter values to local variables at start of script.
This change is part of a booking keeping effort to attempt to get all tunable parameters for the pipeline into a single configuration file.   We'll need to make some more of the parameters set in the fortran and script codes accessible from these drivers before we attain that goal.  baby steps.