Changed the mag range for bright planted sources.
iources are added to the images in 3 groups.
Bright sources, Slow faint source, Fast faint sources.
The Birght source range was set to be 21.0 to 23.5 but the OSSOS
observations typically start having flux and centroid problems
around 23.0.  The bight sources are used as a check of the
flux calibration so we need them to be bright enough to make good
flux measurements. astrom_mag_check.py sets this boundary
at 23.0 so putting it as 22.8 here ensures that the there are
enough bright sources for calibration.