Corrected the threshold where flipping occurs.
The ccds are numbered as ccd = extno-1,  The rule for flipping in storage is based on extno which is <19
but here we are deciding to flip based on ccdno so need to branch at ccd < 18.  The logic for flipping should
really be in one place and should not require this hard-coded stuff.  See the TODO.