Make the storage of results to vospace more robust and push to vospace even on exceptions.
The previous approach was to open a file for writing directly in VOSpace and then only write
the results if the number of bright objects detected was sufficiently high.  This made failures
frequent (VOSpace connection drops) and also made failures hard to diagnose (no results
written on exception).  The new approach is easier to diagnose issues.