stopped flipping locally stored files.
The get_image method will return a local copy of a file when possible (other change in this commit). This
allows the astrometric transform appraoch to based on a locally stored copy of th image an so flip/flopping
is handled elsewhere, nominally.

This is still fragile. There are cases when there isn't a local copy available and then we need to correctly
flip/flop the image.  implementing the fix is a TBD.