Remove the restriction of MERR and raise a more useful message.
The previous daophot would not return a magnutude if MERR was
large. But large is a relative term.  So, now daophot returns
a magnitude if no errors were raised, even if that value has
a large uncertainty. The uncertainty is sent back too, so
the downstream user can decide what to do.

Also, task error message that was sent with errors was not
very specific. Now its a bit clearer.