Added a new OSSOSComment class.
The comment from string method needed a pretty major overhaul.  The motivation for the changes was to add
some extra information to the comment lines that can be used by the TNOdb.  Specifically the astrometric
level (which is really the calibration level). Old style comments are appear to be handled correctly still.

I've not written the unit tests that we need to have
to make sure that future mods to OSSOSComment don't break what we have now.