Made the uncertainty estimate on a measurement an input variable.
The OSSOSComment field now has the correct Astrometric Precision set in the line, so we should use that when passing
into librobfit. The drop-back default for lines that don't specify a precision is 0.3'' which was the previous global default.

The global default value is set in the MPC class, not the liborbfit class and can be over-ridden when measurements are loaded.