Corrected the placement of names in record when longer than 7 characters (again)
Some how this change keeps getting reverted or something?  I'm not sure. The format is supposed to be:

Name starts in column 6 unless its longer than 7 characters and then it starts in column 2.  column 1 is
always reserved for the 'null' flag.  The null_observation will return the null_character or a blank when not null.