Turn MPCReader in a reader factory while preserving previous behaviour. Made provisional_name a property.
The previous MPCReader class read the file directly rather than providing a 'read' method.  This is
not the normal behaviour for reader classes.  Now MPCReader initializes the reader properties
and provides a 'read' method can be used to read the file.  If a filename is given when the MPCReader
is created then that file is read into mpc_observations, preserving the old behaviour.

Also, MPCReader now has a 'replace_provisional' argument on initialization.  If this is set then, as
the MPC records are built the provisional_name is replaced with the name of the file that the line is
coming from.