Made selection of psf stars more stable.
The PSF is built from a list of stars provided by step0jmp.  This list sometimes has
galaxies in it. The CHI cut on output of nstars is dynamically tuned to slowly
remove the things that are mostly likley not stars.  The base assumption is that
most of things going into the PSF are stars.