Moved to using direct imports of numpy methods, also create a 'main' function to avoid variable shadowing.
The PyCharm editor doesn't resolve numpy.radians() correctly, but if one does "from numpy import radians" then the editor
does know what to do.  This make the highlighting of problem areas in the code and revealed, in this instance, that the
exit code was not being set correctly on failures.