Change Range to accept a tuple as well as two distinct arguments for min,max range.  Also added a class method does does the real work.
KBOGenerator is a class but the real use is to call the method get_kbos which instantiates a new instance of this class and then
returns the iterated list of KBOs that the requested ranges would produce.  A bit clunky but likely the correct way to handle this sort
of generator object.