Added some warnings messages when zmag appears to be arbitrary or inconsistent.
The exception is that .zeropoint.used should contain the ZP used when planting sources into the image.  The PHOTZP value in
the header, which is updated when better calibrations are available, should be used when measuring real sources.  So, we just check
that they are consistent (which they don't need to be but often are).