Made the mag distribution a step-function with 30 coming from mag<23.3 and 70% from mag>=23.3
The flux distributoin of planted sources needs to be solid enough near the transition to 0 completeness such that we can
measure the shape. Also, we don't want too many bright sources in planted images that the operator looks at since
that will create a distraction.  The solution is to have a step at around 23.3 where we transition from full completeness to
fractional completelness.