Changed where TimeMPC comes from and added support for cutouts content_decomposition.
Storage now supports calls to the RA/DEC cutout service.  This support includes packing the
boundaries of the cutout into the header of the fits file that is returned.

We now have a number of ways of doing cutouts.  This is getting a bit messing and soon we must refactor
the entire image access layer... soon.