Changed logic used to determine extension number when one is not provided.
The readings have an hdulist object and the reading location (reading.x/reading.y etc.)
is for a particular extension in that hdulist.  This value needs to be set, sometimes
this is set by the person building the reading.  If this isn't set then the
WCS to translate the RA/DEC to an X/Y and determine which extension that works for.
If no RA/DEC is given just set extno=1 and hope for the best.