Removed some of the un-used matplotlib display functoins, centralized the connection to ds9.
We moved to using DS9 a few versions back but had not removed the unused matplotlib calls.
For ds9 the varous calls had been inserted into the code where ever handy, now all connections
to ds9 are handled via a connection setup inside the 'imageview' module.

Call to align images now part of the baseviewer which must be initialized with a connection
to the ds9 display.  The baseviewer is created inside the imageview object.  The viewer object
then passes that ds9 connection to the Displayable when an actual display occurs.

Class structure is something like:  ImageView -> SingletViewer -> DisplayableImageSingle