Corrected offset coordinate problem when building cutouts.
The storage module builds the RA/DEC cutout via a direct call to VOSpace.  Part of this call sets the 'offset' between the
original frame and the cutout.  The previous setting added a '-1' to the offset.  But that is not actually needed given how
the offset is used as X_cutout = X_original -  offset.   ie:  offset = X_original - X_cutout  (same for y BTWs:0)