Corrected use of x/y focus location so that image displays align correctly.
There was 'focus' method that was being ignored and did the correct reference system offsets to allign images.
To make use of this method in setting the ds9 RA/DEC alignmnet (as needed for displaying mosaics of FITS HDULISTS)
a new method was added to the baseviewer and this method is now called from the appview after images are loaded.