More corrections to use of x/y focus location so that image displays align correctly.
The previous align method was ignoring the fact that the focus coordinates were coming
from the processing pipeline and so might be flipped in x/y. Now the focus coordinate is
flipped (if need be) prior to being used as a reference.  This process expects the x/y location
in focus to be from the MOP system.