Added class to handle formatting DS9 regions, also added units to the 'Ellipse' object and made Ellpise orientation match ds9 convention.
The DS9 markers are specified as 'regions' that have specifically formatted initialization strings
depending on the shape of region being initialized.  The construction of those format strings
is now done by the Regions class.  The action to placemarkers is handled by the 'viewer' objects
and the actualy placement of the marker is done in a 'displayable' object.