renamed parameter 'all' in ossos_discoveries to all_releases. Also moved some functions from plot_efficiency to utils
the 'all' parameter specifies that objects from all releases should be parsed into the file. The name is easier to understand and doesn't
clash with the built in variable 'all'.

This module set needs some refactoring.  creating 'utils' was needed so that OSSOS_plot, which uses parses, doesn't depend on plot_efficency.