Have tracks controller update the internal copy of the mpc_observation.
The writer class for the tracks workunit does not persist changes made to obsevation records after writing them to a file.
The controller now updates the data in the workunit to ensure that repeated saves of the current work does not caused previously updated observations to revert to those that
were used at the start of the process.