Use the 'usepv' flag to convert between PV RA/DEC WCS and non PV RA/DEC WCS
SGwyn's WCS uses PV keywords that ds9 doesn't understand.  To make a TRUE RA/DEC location on the image that location
must first be converted to pixel wiht the PV keywords and then the pixels convert to RA/DEC without PV values.  Previously
this method created a special header to handle this, now the pix2world routine knows if it should turn off the PV part.