Use pixel coordinates to mark source location when running in cands/reals  else use RA/DEC
The .{cands,reals}.astrom files contain both the RA/DEC and the X/Y location of the source in the validation image.  The RA/DEC
is derived at the time the pipeline ran and the WCS now in the image header may have been updated.  Thus, we need to use the X/Y
location to mark where the source is, as the RA/DEC solution could be changing.