The ssos.Query object now supports query on a list of observations or an object name.
Implementation of this is done by inspecting the the 'type' of variable passed to query.  If the variable is  list or array then we assume this is a set of observations.  If the list is a string then we assume that we have an object name.   Watch out, we are assuming too much.