Substantial changes to the SourceCutout object to remove mulitple storage locations of X/Y and RA/DEC locations.
Each SourceCutout has an assocaited 'reading'.  The reading is where the X/Y and RA/DEC of the measurement of the source (the SourceReading) are stored.
Over time a number of 'convenience' methods were added to SourceCutout to make the position of the source in the cutout available fro the SourceCutout object. This creates some confusion about which values are the current best measures of the X/Y, RA/DEC of the souce.  So, we ahve changed the SourceCutout methods to always revert to the SourceReading when asked what the position of the source in the cutout is.   The only thing the SourceCutout does is to translate the full frame coordinates, as stored in the SourceReading  into the Cutout Frame of Reference.